Games | refactoring all project (#102)
- delete russian words
- refactor project
- divide models
- add db_table in genres
