+++ backend/games/core/models/__init__.py
@@ -1,0 +1,2 @@
+from .product import *  # noqa
+from .other import *  # noqa
