+++ backend/games/reference/models/__init__.py
@@ -1,0 +1,2 @@
+from .langs import *  # noqa
+from .genres import *  # noqa
