+++ backend/games/reference/serializers.py
@@ -3,16 +3,12 @@ from .models import Language, ProductLanguage, Genre, SubGenre
 
 
 class LanguageSerializer(serializers.ModelSerializer):
-    """Язык"""
-
     class Meta:
         model = Language
         fields = ('id', 'name')
 
 
 class ProductLanguageSerializer(serializers.ModelSerializer):
-    """Поддерживаемый язык у игры"""
-
     language_name = serializers.CharField(source='language.name')
 
     class Meta:
@@ -27,24 +23,18 @@ class ProductLanguageSerializer(serializers.ModelSerializer):
 
 
 class GenreSerializer(serializers.ModelSerializer):
-    """ Жанры """
-
     class Meta:
         model = Genre
         exclude = ('products',)
 
 
 class SubGenreSerializer(serializers.ModelSerializer):
-    """Поджанры"""
-
     class Meta:
         model = SubGenre
         fields = ('id', 'name')
 
 
 class GenreGamesSerializer(serializers.ModelSerializer):
-    """Жанры"""
-
     subgenres = SubGenreSerializer(many=True, read_only=True)
 
     class Meta:
